Renamed WPStatsTrackerMixpanelInstructionsForStat.
Renamed WPStatsTrackerMixpanelInstructionsForStat to
AnalyticsTrackerMixpanelInstructionsForStat
